Added comments. TODO:   1. Destructors   2. Tests   3. Readme

+++ numberSwap.h
@@ -3,25 +3,30 @@
 
 #include <fstream>
 #include "cipher.h"
+
 using namespace std;
-// numberSwap structure.
-class NumberSwap: public Cipher {
+
+//------------------------------------------------------------------------------
+// numberSwap.h - Implements cipher of changing letters to number (inherited from cipher).
+//------------------------------------------------------------------------------
+class NumberSwap : public Cipher {
 public:
     virtual ~NumberSwap() {}
-    // Inputting numberSwap.
+
+    // Inputting numberSwap from stream.
     virtual void In(ifstream &stream);
-    // Creating random numberSwap.
+
+    // Generating random numberSwap.
     virtual void InRnd();
-    // Outputting numberSwap.
+
+    // Outputting numberSwap to stream.
     virtual void Out(ofstream &stream);
+
 private:
     short encryptedScript[10000];
-    int table[26];
+    // Table of exchanging (i's letter of alphabet to i's short)
+    short table[26];
 };
-
-
-
-
 
 
 #endif // __numberSwap__
